+++ eventsk1.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Workspace
+   version = "1.0">
+   <FileRef
+      location = "self:/Users/cec804/Box Sync/ORG-SCHOOL-WCAS-LINGUISTICS-WELLWOOD-LAB/project: events/experiments/Eventsk1v1 2016-08-10 13-27-15/eventsk1_dev/eventsk1.xcodeproj">
+   </FileRef>
+</Workspace>
